added the measure taker
